refactor: Avoid unnecessary monomorphizations
Removes ~80kb which isn't as much as I had hoped for, there are still
some redundant monomorphizations but those are trickier or have some
performance tradeoffs.
