+++ vm/src/core/mod.rs
@@ -30,6 +30,8 @@ macro_rules! assert_deq {
 #[cfg(test)]
 #[cfg_attr(rustfmt, rustfmt_skip)]
 mod grammar;
+#[cfg(test)]
+mod pretty;
 pub mod optimize;
 pub mod interpreter;
 
@@ -44,8 +46,6 @@ use itertools::Itertools;
 use self::typed_arena::Arena;
 
 use self::smallvec::SmallVec;
-
-use pretty::{self, DocAllocator};
 
 use base::ast::{self, Literal, SpannedExpr, SpannedPattern, Typed, TypedIdent};
 use base::fnv::FnvSet;
@@ -101,17 +101,28 @@ pub enum Expr<'a> {
     Match(&'a Expr<'a>, &'a [Alternative<'a>]),
 }
 
+#[cfg(test)]
 impl fmt::Display for Pattern {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        use pretty;
         let arena = pretty::Arena::new();
         let mut s = Vec::new();
         self.pretty(&arena).1.render(80, &mut s).unwrap();
         write!(f, "{}", ::std::str::from_utf8(&s).expect("utf-8"))
     }
 }
+#[cfg(not(test))]
+impl fmt::Display for Pattern {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
 
+#[cfg(test)]
 impl<'a> fmt::Display for Expr<'a> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        use core::pretty::Prec;
+        use pretty;
         let arena = pretty::Arena::new();
         let mut s = Vec::new();
         self.pretty(&arena, Prec::Top).1.render(80, &mut s).unwrap();
@@ -119,7 +130,12 @@ impl<'a> fmt::Display for Expr<'a> {
     }
 }
 
-const INDENT: usize = 4;
+#[cfg(not(test))]
+impl<'a> fmt::Display for Expr<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
 
 #[derive(Default)]
 #[must_use]
@@ -161,165 +177,7 @@ impl<'a> Binder<'a> {
     }
 }
 
-#[derive(Clone, Copy)]
-pub enum Prec {
-    Top,
-    Arg,
-}
-
-impl Prec {
-    pub fn enclose<'a>(
-        &self,
-        arena: &'a pretty::Arena<'a>,
-        doc: pretty::DocBuilder<'a, pretty::Arena<'a>>,
-    ) -> pretty::DocBuilder<'a, pretty::Arena<'a>> {
-        if let Prec::Arg = *self {
-            chain![arena; "(", doc, ")"]
-        } else {
-            doc
-        }
-    }
-}
-
-fn pretty_literal<'a>(
-    l: &Literal,
-    arena: &'a pretty::Arena<'a>,
-) -> pretty::DocBuilder<'a, pretty::Arena<'a>> {
-    match *l {
-        Literal::Byte(b) => arena.text(format!("b{}", b)),
-        Literal::Char(c) => arena.text(format!("{:?}", c)),
-        Literal::Float(f) => arena.text(format!("{}", f)),
-        Literal::Int(i) => arena.text(format!("{}", i)),
-        Literal::String(ref s) => arena.text(format!("{:?}", s)),
-    }
-}
-
 impl<'a> Expr<'a> {
-    pub fn pretty(
-        &'a self,
-        arena: &'a pretty::Arena<'a>,
-        prec: Prec,
-    ) -> pretty::DocBuilder<'a, pretty::Arena<'a>> {
-        match *self {
-            Expr::Call(f, args) => {
-                let doc = chain![arena;
-                    f.pretty(arena, Prec::Arg),
-                    arena.concat(args.iter().map(|arg| {
-                        arena.space().append(arg.pretty(arena, Prec::Arg))
-                    }))
-                ].group();
-                prec.enclose(arena, doc)
-            }
-            Expr::Const(ref l, _) => pretty_literal(l, arena),
-            Expr::Data(ref ctor, args, ..) => match *ctor.typ {
-                Type::Record(ref record) => chain![arena;
-                            "{",
-                            arena.space(),
-                            arena.concat(record.row_iter().zip(args).map(|(field, arg)| {
-                                chain![arena;
-                                    field.name.as_ref(),
-                                    " =",
-                                    arena.space(),
-                                    arg.pretty(arena, Prec::Top),
-                                    ","
-                                ]
-                            })),
-                            arena.space(),
-                            "}"
-                        ].group(),
-                _ => {
-                    let doc = chain![arena;
-                            ctor.as_ref(),
-                            arena.concat(args.iter().map(|arg| {
-                                arena.space().append(arg.pretty(arena, Prec::Arg))
-                            }))
-                        ].group();
-                    prec.enclose(arena, doc)
-                }
-            },
-            Expr::Ident(ref id, _) => arena.text(id.as_ref()),
-            Expr::Let(ref bind, ref expr) => {
-                let doc = chain![arena;
-                    "let ",
-                    match bind.expr {
-                        Named::Expr(ref expr) => {
-                            chain![arena;
-                                bind.name.as_ref(),
-                                arena.space(),
-                                "=",
-                                arena.space(),
-                                chain![arena;
-                                    expr.pretty(arena, Prec::Top),
-                                    arena.space()
-                                ].group()
-                            ].group().nest(INDENT)
-                        }
-                        Named::Recursive(ref closures) => {
-                            arena.concat(closures.iter().map(|closure| {
-                                chain![arena;
-                                    closure.name.as_ref(),
-                                    arena.concat(closure.args.iter()
-                                        .map(|arg| arena.space().append(arena.text(arg.as_ref())))),
-                                    arena.space(),
-                                    "=",
-                                    arena.space(),
-                                    chain![arena;
-                                        closure.expr.pretty(arena, Prec::Top),
-                                        arena.space()
-                                    ].nest(INDENT).group()
-                                ].group()
-                            }))
-                        }
-                    },
-                    arena.newline(),
-                    expr.pretty(arena, Prec::Top)
-                ];
-                prec.enclose(arena, doc)
-            }
-            Expr::Match(expr, alts) => match alts.first() {
-                Some(
-                    alt @ &Alternative {
-                        pattern: Pattern::Record(..),
-                        ..
-                    },
-                ) if alts.len() == 1 =>
-                {
-                    let doc = chain![arena;
-                        "match ",
-                        expr.pretty(arena, Prec::Top),
-                        " with",
-                        arena.newline(),
-                        chain![arena;
-                            alt.pattern.pretty(arena),
-                            arena.space(),
-                            "->"
-                        ].group(),
-                        arena.newline(),
-                        alt.expr.pretty(arena, Prec::Top).group()
-                    ].group();
-                    prec.enclose(arena, doc)
-                }
-                _ => {
-                    let doc = chain![arena;
-                        "match ",
-                        expr.pretty(arena, Prec::Top),
-                        " with",
-                        arena.newline(),
-                        arena.concat(alts.iter().map(|alt| {
-                            chain![arena;
-                                alt.pattern.pretty(arena),
-                                " ->",
-                                arena.space(),
-                                alt.expr.pretty(arena, Prec::Top).nest(INDENT).group()
-                            ].nest(INDENT)
-                        }).intersperse(arena.newline()))
-                    ].group();
-                    prec.enclose(arena, doc)
-                }
-            },
-        }
-    }
-
     pub fn span(&self) -> Span<BytePos> {
         match *self {
             Expr::Call(expr, args) => {
@@ -340,45 +198,6 @@ impl<'a> Expr<'a> {
                 let span_start = expr.span();
                 Span::new(span_start.start(), alts.last().unwrap().expr.span().end())
             }
-        }
-    }
-}
-
-impl Pattern {
-    pub fn pretty<'a>(
-        &'a self,
-        arena: &'a pretty::Arena<'a>,
-    ) -> pretty::DocBuilder<'a, pretty::Arena<'a>> {
-        match *self {
-            Pattern::Constructor(ref ctor, ref args) => chain![arena;
-                    ctor.as_ref(),
-                    arena.concat(args.iter().map(|arg| {
-                        arena.space().append(arg.as_ref())
-                    }))
-                ].group(),
-            Pattern::Ident(ref id) => arena.text(id.as_ref()),
-            Pattern::Record(ref fields) => chain![arena;
-                    "{",
-                    arena.concat(fields.iter().map(|&(ref field, ref value)| {
-                        chain![arena;
-                            arena.space(),
-                            arena.text(field.as_ref()),
-                            match *value {
-                                Some(ref value) => {
-                                    chain![arena;
-                                        "=",
-                                        arena.space(),
-                                        value.as_ref()
-                                    ]
-                                }
-                                None => arena.nil(),
-                            }
-                        ]
-                    }).intersperse(arena.text(","))).nest(INDENT),
-                    arena.space(),
-                    "}"
-                ].group(),
-            Pattern::Literal(ref l) => pretty_literal(l, arena),
         }
     }
 }
@@ -1578,6 +1397,13 @@ impl<'a, 'e> PatternTranslator<'a, 'e> {
     where
         I: IntoIterator<Item = &'b SpannedPattern<Symbol>>,
     {
+        self.pattern_identifiers_(&mut patterns.into_iter())
+    }
+
+    fn pattern_identifiers_<'b, 'p: 'b>(
+        &self,
+        patterns: &mut Iterator<Item = &'b SpannedPattern<Symbol>>,
+    ) -> (Pattern, HashMap<Symbol, Symbol>) {
         let mut identifiers: Vec<TypedIdent<Symbol>> = Vec::new();
         let mut record_fields: Vec<(TypedIdent<Symbol>, _)> = Vec::new();
         let mut core_pattern = None;
